refactor(core,db,connectors,server): code health cleanup (-3,964 LOC)
Extract sql_analysis module from stream_executor, remove duplicate
utility functions, introduce CheckpointRequest parameter struct,
feature-gate delta mode, clean up unused dependencies, deduplicate
cross-crate constants, fix doc inaccuracies, and improve error chain
preservation.

Key changes:

- Extract 16 utility functions from stream_executor.rs into new
  sql_analysis.rs module; remove 1,371 LOC of duplicate function
  bodies from stream_executor, update all import sites
- Introduce CheckpointRequest struct replacing 5-7 positional HashMap
  parameters on checkpoint methods; remove clippy::too_many_arguments
  suppressions
- Gate delta/cluster server mode behind `delta-experimental` feature
  flag with clear "not production-ready" error when disabled
- Remove unused deps: papaya, bumpalo, crc32c, num_cpus, anyhow,
  windows-sys, base64, laminar-storage (phantom dep in connectors)
- Feature-gate xorf behind `files` feature, fix criterion workspace
  version conflict (0.5 vs 0.8), update all inter-crate version
  pins from 0.18.10 to 0.18.13
- Move JSONB binary format tags to laminar-core, both crates re-export;
  consolidate error codes (LDB-1xxx..3xxx) into laminar-core with
  laminar-sql re-exporting via `pub use`
- Add structured DbError variants (CheckpointStore, ConnectorOp) with
  #[from] wrappers preserving error source chains and is_transient()
  delegation
- Fix sink write_to_sinks to break on error instead of continuing;
  add events_dropped counter + error! logging for table upsert failures
- Extract DDL connector resolution into resolve_connector_info() and
  validate_format(); extract operator_graph helpers
  (ensure_query_source_nodes, wire_query_edges, finish_cycle)
- Rename operators/ -> operator/ for consistency with laminar-core;
  rename CDC metrics/registration functions for cross-connector
  consistency; make MySQL CDC submodules pub mod
- Fix stale docs: window.rs "(future)" markers, misplaced
  handle_restore_checkpoint docstring, CheckpointStore::load_by_id
  contract, source() error doc, SourceCheckpoint::is_empty(),
  WsBackpressure unimplemented variant docs, EmitStrategy verbose
  variant enumeration
