+++ .gitignore
@@ -57,3 +57,4 @@ docs/features/
 docs/research/
 docs/audits/
 docs/adr/
+.desloppify/
