+++ crates/laminar-connectors/Cargo.toml
@@ -11,8 +11,7 @@ description = "External system connectors for LaminarDB - Kafka, CDC, lookup tab
 
 [dependencies]
 # Core dependencies
-laminar-core = { path = "../laminar-core", version = "0.18.10" }
-laminar-storage = { path = "../laminar-storage", version = "0.18.10" }
+laminar-core = { path = "../laminar-core", version = "0.18.13" }
 
 # Kafka
 rdkafka = { version = "0.39", features = ["cmake-build", "ssl-vendored"], optional = true }
@@ -86,8 +85,7 @@ futures-util = { version = "0.3", features = ["sink"], optional = true }
 # File source/sink
 globset = { workspace = true, optional = true }
 notify = { workspace = true, optional = true }
-xorf = { workspace = true }
-base64 = { workspace = true }
+xorf = { workspace = true, optional = true }
 libc = { workspace = true, optional = true }
 
 [features]
@@ -116,14 +114,14 @@ kafka-discovery = ["kafka", "laminar-core/delta", "dep:tokio-util"]
 # WebSocket source and sink
 websocket = ["dep:tokio-tungstenite", "dep:tungstenite", "dep:futures-util", "dep:url"]
 # File source and sink (AutoLoader-style)
-files = ["dep:notify", "dep:globset", "dep:object_store", "dep:parquet", "dep:libc", "dep:tokio-stream"]
+files = ["dep:notify", "dep:globset", "dep:object_store", "dep:parquet", "dep:libc", "dep:tokio-stream", "dep:xorf"]
 
 [dev-dependencies]
 tempfile = { workspace = true }
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 testcontainers = "0.27"
 testcontainers-modules = { version = "0.15", features = ["postgres"] }
-criterion = { version = "0.5", features = ["html_reports"] }
+criterion = { workspace = true }
 
 [[bench]]
 name = "mongodb_throughput"
