+++ crates/laminar-core/Cargo.toml
@@ -23,14 +23,12 @@ delta = [
 # Performance critical - no allocations on hot path
 rustc-hash = { workspace = true }
 ahash = { workspace = true }
-papaya = { workspace = true }
 xorf = { workspace = true }
 parking_lot = { workspace = true }
 bytes = { workspace = true }
 foyer = { workspace = true }
 equivalent = "1"
 memmap2 = { workspace = true }
-bumpalo = { workspace = true }
 uuid = { workspace = true }
 smallvec = { workspace = true }
 arrayvec = { workspace = true }
@@ -43,12 +41,8 @@ arrow-cast = { workspace = true }
 arrow-ipc = { workspace = true }
 chrono = { workspace = true }
 
-# Integrity
-crc32c = "0.6"
-
 # Error handling
 thiserror = { workspace = true }
-anyhow = { workspace = true }
 
 # Serialization for state (rkyv for zero-copy, serde for compatibility)
 rkyv = { workspace = true }
@@ -72,9 +66,6 @@ xxhash-rust = { workspace = true, optional = true }
 # Observability
 tracing = { workspace = true }
 
-# Thread-per-core
-num_cpus = "1.17"
-
 # NUMA topology (optional, enhanced detection)
 hwlocality = { version = "1.0.0-alpha.8", optional = true }
 
@@ -85,9 +76,6 @@ io-uring = { workspace = true, optional = true }
 
 [target.'cfg(target_os = "macos")'.dependencies]
 libc = { workspace = true }
-
-[target.'cfg(target_os = "windows")'.dependencies]
-windows-sys = { version = "0.59", features = ["Win32_System_Threading"] }
 
 [dev-dependencies]
 criterion = { workspace = true }
