+++ crates/laminar-db/Cargo.toml
@@ -36,10 +36,10 @@ gcs = ["laminar-storage/gcs"]
 azure = ["laminar-storage/azure"]
 
 [dependencies]
-laminar-core = { path = "../laminar-core", version = "0.18.10" }
-laminar-sql = { path = "../laminar-sql", version = "0.18.10" }
-laminar-storage = { path = "../laminar-storage", version = "0.18.10" }
-laminar-connectors = { path = "../laminar-connectors", version = "0.18.10" }
+laminar-core = { path = "../laminar-core", version = "0.18.13" }
+laminar-sql = { path = "../laminar-sql", version = "0.18.13" }
+laminar-storage = { path = "../laminar-storage", version = "0.18.13" }
+laminar-connectors = { path = "../laminar-connectors", version = "0.18.13" }
 
 # Arrow
 arrow = { workspace = true }
@@ -52,16 +52,12 @@ datafusion = { workspace = true }
 datafusion-common = { workspace = true }
 datafusion-expr = { workspace = true }
 
-# Arena allocator
-bumpalo = { workspace = true }
-
 # Serialization
 serde = { workspace = true }
 serde_json = { workspace = true }
 
 # Error handling
 thiserror = { workspace = true }
-anyhow = { workspace = true }
 
 # SQL parser (for AST type references)
 sqlparser = { workspace = true }
@@ -90,8 +86,8 @@ tracing = { workspace = true }
 [dev-dependencies]
 tokio = { workspace = true, features = ["test-util", "macros"] }
 tempfile = { workspace = true }
-laminar-derive = { path = "../laminar-derive", version = "0.18.10" }
-laminar-connectors = { path = "../laminar-connectors", version = "0.18.10", features = ["testing"] }
+laminar-derive = { path = "../laminar-derive", version = "0.18.13" }
+laminar-connectors = { path = "../laminar-connectors", version = "0.18.13", features = ["testing"] }
 criterion = { workspace = true }
 
 [[bench]]
