+++ crates/laminar-server/src/server.rs
@@ -17,6 +17,7 @@ use laminar_db::{DbError, LaminarDB, Profile};
 use crate::config::{
     ConfigError, LookupConfig, PipelineConfig, ServerConfig, SinkConfig, SourceConfig,
 };
+#[cfg(feature = "delta-experimental")]
 use crate::delta_config::{DeltaConfig, DeltaConfigError};
 use crate::http;
 use crate::reload::ReloadGuard;
@@ -36,6 +37,10 @@ pub enum ServerHandle {
         watcher_handle: Option<tokio::task::JoinHandle<()>>,
     },
     /// Delta (multi-node) mode.
+    ///
+    /// Only available when the `delta-experimental` feature is enabled.
+    /// This mode is not yet production-ready.
+    #[cfg(feature = "delta-experimental")]
     Delta(Box<crate::delta::DeltaHandle>),
 }
 
@@ -65,6 +70,7 @@ impl ServerHandle {
                 info!("Shutdown complete");
                 Ok(())
             }
+            #[cfg(feature = "delta-experimental")]
             Self::Delta(handle) => (*handle)
                 .wait_for_shutdown()
                 .await
@@ -111,13 +117,25 @@ pub async fn run_server(
     config: ServerConfig,
     config_path: PathBuf,
 ) -> Result<ServerHandle, ServerError> {
-    let delta_cfg = DeltaConfig::from_server_config(&config)?;
+    // Delta mode: gate behind the `delta-experimental` feature flag.
+    #[cfg(feature = "delta-experimental")]
+    {
+        let delta_cfg = DeltaConfig::from_server_config(&config)?;
 
-    if let Some(delta_cfg) = delta_cfg {
-        let handle = crate::delta::start_delta(config, delta_cfg, config_path)
-            .await
-            .map_err(|e| ServerError::Delta(e.to_string()))?;
-        return Ok(ServerHandle::Delta(Box::new(handle)));
+        if let Some(delta_cfg) = delta_cfg {
+            let handle = crate::delta::start_delta(config, delta_cfg, config_path)
+                .await
+                .map_err(|e| ServerError::Delta(e.to_string()))?;
+            return Ok(ServerHandle::Delta(Box::new(handle)));
+        }
+    }
+    #[cfg(not(feature = "delta-experimental"))]
+    if config.server.mode == "delta" {
+        return Err(ServerError::Delta(
+            "Delta/cluster mode requires the 'delta-experimental' feature flag. \
+             This mode is not yet production-ready."
+                .to_string(),
+        ));
     }
 
     // 1. Build LaminarDB via builder API
@@ -489,6 +507,7 @@ pub enum ServerError {
     Delta(String),
 
     /// Delta configuration error.
+    #[cfg(feature = "delta-experimental")]
     #[error(transparent)]
     DeltaConfig(#[from] DeltaConfigError),
 }
