+++ crates/laminar-sql/Cargo.toml
@@ -11,7 +11,7 @@ description = "SQL layer for LaminarDB with streaming extensions"
 
 [dependencies]
 # Core dependency
-laminar-core = { path = "../laminar-core", version = "0.18.10" }
+laminar-core = { path = "../laminar-core", version = "0.18.13" }
 
 # DataFusion for SQL processing
 datafusion = { workspace = true }
@@ -30,12 +30,11 @@ sqlparser = { workspace = true }
 
 # Error handling
 thiserror = { workspace = true }
-anyhow = { workspace = true }
 
 # Async
 tokio = { workspace = true }
 async-trait = { workspace = true }
-futures = "0.3"
+futures = { workspace = true }
 
 # Concurrency
 parking_lot = { workspace = true }
