+++ examples/binance-ws/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "binance-ws"
-version = "0.18.12"
+version = "0.18.13"
 edition = "2021"
 publish = false
 description = "Binance VWAP signals: minimal LaminarDB streaming example"
@@ -10,8 +10,8 @@ name = "binance-ws"
 path = "src/main.rs"
 
 [dependencies]
-laminar-db = { path = "../../crates/laminar-db", version = "0.18.12", features = ["websocket"] }
-laminar-derive = { path = "../../crates/laminar-derive", version = "0.18.12" }
+laminar-db = { path = "../../crates/laminar-db", version = "0.18.13", features = ["websocket"] }
+laminar-derive = { path = "../../crates/laminar-derive", version = "0.18.13" }
 arrow = "57.2"
 tokio = { version = "1.50", features = ["full"] }
 crossterm = "0.29"
