+++ examples/demo/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "laminardb-demo"
-version = "0.18.12"
+version = "0.18.13"
 edition = "2021"
 publish = false
 description = "Market Data Demo: Real-time analytics with Ratatui TUI for LaminarDB"
@@ -23,9 +23,9 @@ default = []
 kafka = ["laminar-db/kafka", "dep:rdkafka", "dep:serde", "dep:serde_json"]
 
 [dependencies]
-laminar-db = { path = "../../crates/laminar-db", version = "0.18.12" }
-laminar-derive = { path = "../../crates/laminar-derive", version = "0.18.12" }
-laminar-core = { path = "../../crates/laminar-core", version = "0.18.12" }
+laminar-db = { path = "../../crates/laminar-db", version = "0.18.13" }
+laminar-derive = { path = "../../crates/laminar-derive", version = "0.18.13" }
+laminar-core = { path = "../../crates/laminar-core", version = "0.18.13" }
 
 # Arrow
 arrow = "57.2"
