+++ examples/microstructure/Cargo.toml
@@ -2,7 +2,7 @@
 
 [package]
 name = "microstructure"
-version = "0.18.12"
+version = "0.18.13"
 edition = "2021"
 publish = false
 description = "Market Microstructure Intelligence: 12 WebSocket streams, 36 SQL stages, sub-us latency"
@@ -12,8 +12,8 @@ name = "microstructure"
 path = "src/main.rs"
 
 [dependencies]
-laminar-db = { path = "../../crates/laminar-db", version = "0.18.12", features = ["websocket"] }
-laminar-derive = { path = "../../crates/laminar-derive", version = "0.18.12" }
+laminar-db = { path = "../../crates/laminar-db", version = "0.18.13", features = ["websocket"] }
+laminar-derive = { path = "../../crates/laminar-derive", version = "0.18.13" }
 arrow = "57.2"
 tokio = { version = "1.50", features = ["full"] }
 crossterm = "0.29"
